#118 n2e_OpenMRULast() moved to extension files
